Rename TypeDescriptor to Converter.
