+++ common/src/main/kotlin/com/github/fkrull/konflin/ConfigSpec.kt
@@ -1,8 +1,8 @@
 package com.github.fkrull.konflin
 
 import com.github.fkrull.konflin.impl.ConverterBasedSetting
-import com.github.fkrull.konflin.typedescriptors.ConfigType
-import com.github.fkrull.konflin.typedescriptors.impl.IdentityTypeDescriptor
+import com.github.fkrull.konflin.converter.ConfigType
+import com.github.fkrull.konflin.converter.impl.IdentityConverter
 import kotlin.reflect.KClass
 
 interface ConfigSpec {
@@ -12,9 +12,9 @@ interface ConfigSpec {
         name: String,
         defaultValue: T? = null
     ): Setting<T> = when (type) {
-        String::class -> ConverterBasedSetting<String>(name, defaultValue as String?, IdentityTypeDescriptor(ConfigType.Types.String))
-        Int::class -> ConverterBasedSetting<Int>(name, defaultValue as Int?, IdentityTypeDescriptor(ConfigType.Types.Int))
-        Boolean::class -> ConverterBasedSetting<Boolean>(name, defaultValue as Boolean?, IdentityTypeDescriptor(ConfigType.Types.Boolean))
+        String::class -> ConverterBasedSetting<String>(name, defaultValue as String?, IdentityConverter(ConfigType.Types.String))
+        Int::class -> ConverterBasedSetting<Int>(name, defaultValue as Int?, IdentityConverter(ConfigType.Types.Int))
+        Boolean::class -> ConverterBasedSetting<Boolean>(name, defaultValue as Boolean?, IdentityConverter(ConfigType.Types.Boolean))
         else -> throw UnsupportedConfigTypeException("cannot use '$type' as config type")
     } as Setting<T>
 }
