+++ common/src/main/kotlin/com/github/fkrull/konflin/impl/ConverterBasedSetting.kt
@@ -2,16 +2,16 @@ package com.github.fkrull.konflin.impl
 
 import com.github.fkrull.konflin.ConfigurationSource
 import com.github.fkrull.konflin.Setting
-import com.github.fkrull.konflin.typedescriptors.ConfigType
-import com.github.fkrull.konflin.typedescriptors.TypeDescriptor
+import com.github.fkrull.konflin.converter.ConfigType
+import com.github.fkrull.konflin.converter.Converter
 
 internal class ConverterBasedSetting<out T : Any>(
     override val name: String,
     override val default: T?,
-    converter: TypeDescriptor<T, *>
+    converter: Converter<T, *>
 ) : Setting<T> {
     @Suppress("UNCHECKED_CAST")
-    private val converter: TypeDescriptor<T, Any> = converter as TypeDescriptor<T, Any>
+    private val converter: Converter<T, Any> = converter as Converter<T, Any>
 
     override fun get(configSource: ConfigurationSource): T? {
         return when (converter.configType) {
